Broken ts commands

+++ ULSS Helper/Objects/TS.cs
@@ -2,7 +2,7 @@
 
 public class TS
 {
-    public ulong ID { get; set; }
+    public int ID { get; set; }
     public string Username { get; set; }
     public int View { get; set; }
     public int Allow { get; set; }
